Extract LinkedList as common class
This extract LinkedList and Node as common classes
before writing Stack and Queue, which will use LinkedList.
